+++ README.md
@@ -1,0 +1,22 @@
+# DJ_YAKAMOCHI(Alpha)
+
+様々なサービスの曲を横断して流せる音楽Bot
+
+
+# 機能
+- 音楽の単体再生（YoutubeDLが対応するリソース）
+- プレイリストのランダム・リピート再生（現在Youtubeのみ）
+- DJ再生機能（DJのように曲をシームレスに繋ぐ機能）
+
+# 動作環境（サーバーサイド）
+OS: Linux, Win, Mac
+
+必須パッケージ（パスが通っている必要がある）
+- Java11
+- YoutubeDL
+
+# 動作環境（クライアント）
+- Discord(All Platforms)
+
+予定
+- WebPlayer
