+++ src/main/kotlin/space/siy/dj/yakamochi/App.kt
@@ -2,16 +2,15 @@ package space.siy.dj.yakamochi
 
 import io.ktor.locations.KtorExperimentalLocationsAPI
 import io.ktor.util.KtorExperimentalAPI
-import net.dv8tion.jda.api.JDABuilder
 import org.bytedeco.javacv.FFmpegLogCallback
 import org.jetbrains.exposed.sql.Database
 import org.koin.core.context.startKoin
 import org.koin.dsl.module
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import space.siy.dj.yakamochi.auth.AuthHandler
+import space.siy.dj.yakamochi.gateway.http.AuthHandler
 import space.siy.dj.yakamochi.database.*
-import space.siy.dj.yakamochi.http.startHttpServer
+import space.siy.dj.yakamochi.gateway.Gateways
 
 val repositoryModule = module {
     single { ExposedTrackHistoryRepository() as TrackHistoryRepository }
@@ -24,9 +23,16 @@ val repositoryModule = module {
 @KtorExperimentalAPI
 @ExperimentalStdlibApi
 fun main(args: Array<String>) {
-    System.setProperty("org.bytedeco.javacpp.logger", "slf4j")
-    FFmpegLogCallback.set()
-    startHttpServer()
+    // Logging FFmpeg via slf4j
+    setupFFmpegLogger()
+
+    Database.connect(
+            "jdbc:postgresql://${System.getenv("DB_HOST")}:${System.getenv("DB_PORT")}/${System.getenv("DB_NAME")}",
+            driver = "org.postgresql.Driver",
+            user = System.getenv("DB_USER"),
+            password = System.getenv("DB_PASSWORD")
+    )
+
     startKoin {
         printLogger()
         modules(repositoryModule)
@@ -34,12 +40,13 @@ fun main(args: Array<String>) {
             single { AuthHandler as space.siy.dj.yakamochi.auth.AuthProvider }
         })
     }
-    Database.connect("jdbc:postgresql://${System.getenv("DB_HOST")}:${System.getenv("DB_PORT")}/${System.getenv("DB_NAME")}", driver = "org.postgresql.Driver",
-            user = System.getenv("DB_USER"), password = System.getenv("DB_PASSWORD"))
-    val jda = JDABuilder.createDefault(System.getenv("token")).build()
-    jda.addEventListener(space.siy.dj.yakamochi.music.EventHandler())
-//    jda.addEventListener(DMHandler())
+
+    Gateways.start()
 }
 
+fun setupFFmpegLogger() {
+    System.setProperty("org.bytedeco.javacpp.logger", "slf4j")
+    FFmpegLogCallback.set()
+}
 
 fun Any.logger(): Logger = LoggerFactory.getLogger(this::class.java)
