+++ src/main/kotlin/space/siy/dj/yakamochi/music/track/Track.kt
@@ -13,8 +13,8 @@ import space.siy.dj.yakamochi.music.VideoInfo
 import space.siy.dj.yakamochi.music.audio.AudioProvider
 import space.siy.dj.yakamochi.music.remote.RemoteAudioProvider
 import space.siy.dj.yakamochi.music.sampleCountToSec
-import space.siy.dj.yakamochi.music_service.MusicService
-import space.siy.dj.yakamochi.music_service.MusicServiceManager
+import space.siy.dj.yakamochi.music.service.MusicService
+import space.siy.dj.yakamochi.music.service.MusicServiceManager
 
 /**
  * @author SIY1121
