+++ src/main/kotlin/space/siy/dj/yakamochi/music/track/TrackProvider.kt
@@ -3,7 +3,7 @@ package space.siy.dj.yakamochi.music.track
 import space.siy.dj.yakamochi.Outcome
 import space.siy.dj.yakamochi.music.audio.AudioProvider
 import space.siy.dj.yakamochi.music.remote.RemoteAudioProvider
-import space.siy.dj.yakamochi.music_service.MusicService
+import space.siy.dj.yakamochi.music.service.MusicService
 
 /**
  * @author SIY1121
